requirements.txt
./docs/gradio-3.32.2-py3-none-any.whl
tiktoken>=0.3.3
requests[socks]
transformers
python-markdown-math
beautifulsoup4
prompt_toolkit
latex2mathml
python-docx
mdtex2html
colorama
Markdown
pygments
pymupdf
openai
numpy
arxiv
rich
fastapi[all]
passlib
python-jose[cryptography]